campus-alert-app/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
pillow==10.1.0
eventlet==0.33.3
werkzeug==3.0.1
